[Fact] move Fact to Types folder
