+++ include/procedural/graph/Network.h
@@ -10,7 +10,7 @@
 #include "procedural/graph/Transition.h"
 #include "procedural/graph/State.h"
 #include "procedural/core/FactPattern.h"
-#include "procedural/core/Fact.h"
+#include "procedural/core/Types/Fact.h"
